Moving WorkspaceExecutionPlanner from cli to orchestra module
